Using evdev for mouse (wayland invisible cursor fix) and fixes for 2.0.6 android app protocol
