+++ CMakeLists.txt
@@ -22,6 +22,7 @@ CONFIGURE_FILE(
 
 FIND_PACKAGE(PkgConfig)
 PKG_CHECK_MODULES(GTK gtk+-2.0)
+PKG_CHECK_MODULES(LIBEVDEV libevdev)
 
 # Build of the program
 ADD_EXECUTABLE(${TARGET_NAME} ${SOURCE_FILES})
@@ -34,10 +35,12 @@ TARGET_LINK_LIBRARIES(${TARGET_NAME}
 	pthread
 	avahi-common
 	avahi-client
+	${LIBEVDEV_LIBRARIES}
 	${GTK_LIBRARIES}
 )
 INCLUDE_DIRECTORIES(
 	${CMAKE_CURRENT_BINARY_DIR}
+	${LIBEVDEV_INCLUDE_DIRS}
 	${GTK_INCLUDE_DIRS}
 )
 
@@ -48,7 +51,7 @@ SET(CPACK_DEBIAN_PACKAGE_MAINTAINER "Jim DeVona")
 SET(CPACK_DEBIAN_PACKAGE_HOMEPAGE "https://github.com/anoved/mmserver/")
 SET(CPACK_DEBIAN_PACKAGE_SECTION "X11")
 SET(CPACK_DEBIAN_PACKAGE_DESCRIPTION "Modified Mobile Mouse Server for Linux")
-SET(CPACK_DEBIAN_PACKAGE_DEPENDS "libpcrecpp0, libconfig++8")
+SET(CPACK_DEBIAN_PACKAGE_DEPENDS "libpcrecpp0v5, libconfig++9v5, libevdev2")
 SET(CPACK_DEBIAN_PACKAGE_CONTROL_EXTRA "${CMAKE_SOURCE_DIR}/share/postinst;${CMAKE_SOURCE_DIR}/share/prerm")
 
 # fix permissions of postinst/prerm
