+++ src/server.cpp
@@ -290,7 +290,7 @@ void GTKTrayAbout(GtkMenuItem* item __attribute__((unused)), gpointer uptr __att
 	gtk_about_dialog_set_name((GtkAboutDialog*)about, "Mobile Mouse Server for Linux");
 	gtk_window_set_icon((GtkWindow*)about, gtk_widget_render_icon(about, GTK_STOCK_ABOUT, GTK_ICON_SIZE_MENU, NULL));
 	gtk_about_dialog_set_copyright((GtkAboutDialog*)about, "Copyright (C) 2011 Erik Lax,\n2013 Kiriakos Krastillis, Jim DeVona");
-	gtk_about_dialog_set_version((GtkAboutDialog*)about, MMSERVER_VERSION_MAJOR"."MMSERVER_VERSION_MINOR"."MMSERVER_VERSION_PATCH);
+	gtk_about_dialog_set_version((GtkAboutDialog*)about, MMSERVER_VERSION_MAJOR "." MMSERVER_VERSION_MINOR "." MMSERVER_VERSION_PATCH);
 	gtk_about_dialog_set_website((GtkAboutDialog*)about, "https://github.com/anoved/mmserver/");
 	gtk_about_dialog_set_comments((GtkAboutDialog*)about, "An unofficial Mobile Mouse server for Linux. Use your iOS or Android device as a network mouse and keyboard for your Linux computer using Mobile Mouse and this software.");
 	gtk_about_dialog_set_authors((GtkAboutDialog*)about, authors);
